adding some code

+++ src/main/java/com/codingtech/formationcenter/repo/FormateurRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Formateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormateurRepo  extends JpaRepository<Formateur,Integer> {
+public interface FormateurRepo  extends JpaRepository<Formateur,Long> {
 }
