+++ src/main/java/com/codingtech/formationcenter/repo/ModuleRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Module;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModuleRepo  extends JpaRepository<Module,Integer> {
+public interface ModuleRepo  extends JpaRepository<Module,Long> {
 }
